requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
PyJWT==2.11.0
pyotp==2.9.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
gunicorn
psycopg2-binary
whitenoise
python-dotenv
boto3
django-ses